Some files were added to the public repository to allow compiling the simplelayouts project with Eclipse, so we have to update the additional_files_re to cover these files, so they don't get propogated back to the internal repository.
There was also a use of MoreAsserts in the public repository, some versions of which are open sourced or are part of Android. But the exact version used on client-side is not open sourced, so I replaced it with assertEquals.

For server unit tests, there is an android.test.MoreAsserts class which can probably be used instead.

R=tkaizu,dxu
DELTA=6  (1 added, 0 deleted, 5 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=1934


git-svn-id: http://nativedriver.googlecode.com/svn/trunk@10 9ea60b42-7040-e045-bdf9-16ade6be0113