Causing elements that are in non-focused windows to not appear in findElement(s) results. This fits closer with the philosophy that all Android windows are a single browser window, and each focused Android window represents a different page.
This change was prompted by an issue which occurs when running FindByTextTest on an actual device. (This is the only test that was running more than one test method in a single test class.) When the second test searched for an element, it would often find one in the already-hidden first test's window. That element would be inoperable which would cause the test to fail.

R=dxu,tkaizu
DELTA=208  (158 added, 27 deleted, 23 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2083


git-svn-id: http://nativedriver.googlecode.com/svn/trunk@21 9ea60b42-7040-e045-bdf9-16ade6be0113